requirements.txt
anthropic>=0.40.0
openai>=1.50.0
langdetect>=1.0.9
pandas>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
